apps/scheduler/pyproject.toml
[project]
name = "llms-scheduler"
version = "0.0.0"
description = "Scheduler service for llms.txt generation"
authors = [{name="Your Name", email="you@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "redis >= 4.5.0",
    "python-dotenv >= 1.0.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "requests >= 2.31.0",
]
